fixing norminette and bonus
